refactor(connectors): move full spec into spec.ts default export
Each connector's spec.ts now exports default ConnectorSpecification
(config + stream_state where applicable). The spec() method in each
index.ts yields the default directly instead of inlining z.toJSONSchema
calls. Stripe's stream state Zod schemas move from index.ts to spec.ts.

Committed-By-Agent: claude
